Store new user's password in keyring for Distributed Cloud
In Distributed Cloud when audit tries to create a user missing in subcloud,
the user's password is needed. But keystone API never returns password, so
the password has to be stored in keyring and retrieved from keyring when
needed.

Note that this is a temporary solution for users synchronization in DC.
Once the new synchronization mechanism from Story 2002842 is in place,
we will revisit this change.

Closes-Bug: 1797970

Signed-off-by: Andy Ning <andy.ning@windriver.com>